Function myc::use_cases::default_users::account::create_default_account::create_default_account

pub async fn create_default_account(
    email: String,
    account_name: String,
    first_name: Option<String>,
    last_name: Option<String>,
    user_registration_repo: Box<&dyn UserRegistration>,
    account_type_registration_repo: Box<&dyn AccountTypeRegistration>,
    account_registration_repo: Box<&dyn AccountRegistration>
) -> Result<GetOrCreateResponseKind<AccountDTO>, MappedErrors>

Create a default account.

Default accounts are used to mirror human users. Such accounts should not be flagged as subscription.

This function are called when a new user start into the system. The account-creation method also insert a new user into the database and set the default role as default-user.