Function myc::use_cases::managers::guest_role::create_guest_role::create_guest_role

pub async fn create_guest_role(
    profile: ProfileDTO,
    name: String,
    description: String,
    role: Uuid,
    permissions: Option<Vec<PermissionsType>>,
    role_registration_repo: Box<&dyn GuestRoleRegistration>
) -> Result<GetOrCreateResponseKind<GuestRoleDTO>, MappedErrors>

This function should be called only by manager users. Roles should be created after the application is registered by staff users why roles links guest, permissions, and applications.

As example, a group of users need to has view only permissions to resources of a single application. Thus, the role should include only the View permission (level zero) for the Movie application. Thus, the role name should be: “Movie Viewers”.